server/workers/requirements.txt
colorlog == 2.6.0
pika == 0.10.0
pytz == 2015.7
requests == 2.8.0
semver == 2.2.0
python_dateutil == 2.4.2
redis
numpy
